refactor: supabase/server.ts 위치 변경 (utils)
