+++ src/app/challenges/[id]/page.tsx
@@ -2,8 +2,8 @@ import Image from "next/image"
 import Button from "@/components/common/button/button"
 import ChallengeCardList from "@/components/common/challenge-card-list/challenge-card-list"
 import AvatarLink from "@/components/user/avatar-link/avatar-link"
-import type { Database } from "@/libs/supabase/database.types"
-import { createClient } from "@/libs/supabase/server"
+import type { Database } from "@/utils/libs/supabase/database.types"
+import { createClient } from "@/utils/libs/supabase/server"
 import testImg from "./assets/test.avif"
 import styles from "./page.module.css"
