+++ tsconfig.json
@@ -6,7 +6,7 @@
       "@/components/*": ["src/components/*"],
       "@/styles/*": ["src/styles/*"],
       "@/utils/*": ["src/utils/*"],
-      "@/libs/*": ["src/libs/*"]
+      "@/libs/*": ["src/utils/libs/*"]
     },
     "target": "ES2017",
     "lib": ["dom", "dom.iterable", "esnext"],
